Transform2D fixes.
Transform2D instances now refresh their internal state when added to
and removed from a scene. When a Transform2D is removed from scene it
assumes it's now a top-level transform and recalculates it's members
accordingly. When added to stage, it finds a parent transform and
recalculates from global to this transform's coords.